[505] Add value expression in Node label
Bug: https://github.com/eclipse-syson/syson/issues/505
