+++ CHANGELOG.adoc
@@ -70,6 +70,7 @@ A reset of the database is needed.
 - https://github.com/eclipse-syson/syson/issues/511[#511] [export] Fix an issue where models were exported with a global indentation instead of no indentation. 
 - https://github.com/eclipse-syson/syson/issues/501[#501] [diagrams] Fix an issue where the double quotes were set along with the string value in case of a direct edit of the value part.
 - https://github.com/eclipse-syson/syson/issues/509[#509] [general-view] Add missing doc compartment in UseCase, UseCaseDefinition, and AllocationDefinition
+- https://github.com/eclipse-syson/syson/issues/505[#505] [diagrams] Add value expression in Node label
 
 === Improvements
